More obsolete code removal and renaming
