+++ Sources/PoieticFlows/Simulation/StateVariable.swift
